Move repoConfig to the root of the GitHub folder
